xiaoluzi2.0-projectD/requirements.txt
influxdb_client==1.35.0
keras==2.12.0
numpy==1.23.5
pandas==1.5.1
PyDAQmx==1.4.6
pyserial==3.5
scikit_learn==1.2.2
